+++ CHANGELOG.md
@@ -12,8 +12,8 @@ This project uses a version scheme based on the Cadenza main version in the form
 - `Geometry#coordinates` property
 
 ### Changed
-  - **BREAKING CHANGE** `CadenzaEditGeometryUpdateEvent#detail` is now typed as `FeatureCollection | Feature | undefined`. (Before: `{ geometry: Geometry }`)
-  - `CadenzaEditGeometryOkEvent#detail` is now typed as `Feature` to align it with the rest of the API. (Before: `{ geometry: Geometry }`)
+- **BREAKING CHANGE** `CadenzaEditGeometryUpdateEvent#detail` is now typed as `FeatureCollection | Feature | undefined`. (Before: `{ geometry: Geometry }`)
+- `CadenzaEditGeometryOkEvent#detail` is now typed as `Feature` to align it with the rest of the API. (Before: `{ geometry: Geometry }`)
 
 ## 10.2.2 - 2024-10-24
 ### Added